requirements.txt
streamlit
pandas
pyyaml
matplotlib
google-generativeai
sentence-transformers
torch
plotly
matplotlib